refactor(quantization): simplify SVDQuant integration
